refactor: refactor ccw/ui/terminal_ui.go - large terminal...
- Code refactoring

Resolves #258
